+++ packages/crypto_wallet_util/test/transaction/sc_wasm_asset_test_support_io.dart
@@ -1,0 +1,5 @@
+import 'dart:typed_data';
+
+void configureScWasmAssetForTest(Uint8List wasmBytes) {}
+
+void clearScWasmAssetForTest() {}